Allow configuration from files
Sometimes we don't want to directly specify things like passwords
directly in the environment.  Sometimes, sure, but quite often not.

Here we look for any environment variables named with a `PAPERLESS_`
prefix and `_FROM_FILE` suffix.  If found, we attempt to provision the
value of the root variable (e.g. without the `_FROM_FILE` suffix) from
the contents of the named file.

If the named file doesn't exist, etc, things will explode -- just as
they should.